move and share parse_time function
