+++ graphite-api/src/lib.rs
@@ -3,3 +3,4 @@ pub mod error;
 pub mod find;
 pub mod opts;
 pub mod render;
+pub mod parse;
